ambari stack definition in the format that amabri server expects, config + meta info.
