+++ .idea/.gitignore
@@ -1,0 +1,8 @@
+/.name
+/ambari-zeppelin-service.iml
+/compiler.xml
+/encodings.xml
+/misc.xml
+/modules.xml
+/vcs.xml
+/workspace.xml
