+++ common-services/ZEPPELIN/0.5.5/metainfo.xml
@@ -1,0 +1,51 @@
+<?xml version="1.0"?>
+<metainfo>
+    <schemaVersion>2.0</schemaVersion>
+    <services>
+        <service>
+            <name>ZEPPELIN</name>
+            <displayName>Zeppelin Notebook</displayName>
+            <comment>A web-based notebook that enables interactive data analytics. It enables you to make beautiful data-driven, interactive and collaborative documents with SQL, Scala and more.</comment>
+            <version>0.6.0-SNAPSHOT</version>
+            <components>
+                <component>
+                    <name>ZEPPELIN_MASTER</name>
+                    <displayName>Zeppelin Notebook</displayName>
+                    <category>MASTER</category>
+                    <cardinality>1</cardinality>
+                    <commandScript>
+                        <script>scripts/master.py</script>
+                        <scriptType>PYTHON</scriptType>
+                        <timeout>10000</timeout>
+                    </commandScript>
+                </component>
+                <component>
+                    <name>ZEPPELIN_CLIENT</name>
+                    <displayName>Zeppelin Notebook Client</displayName>
+                    <category>CLIENT</category>
+                    <cardinality>1</cardinality>
+                    <commandScript>
+                        <script>scripts/client.py</script>
+                        <scriptType>PYTHON</scriptType>
+                        <timeout>10000</timeout>
+                    </commandScript>
+                </component>
+            </components>
+            <osSpecifics>
+                <osSpecific>
+                    <osFamily>any</osFamily>
+                    <packages>
+                        <package>
+                            <name>zeppelin</name>
+                        </package>
+                    </packages>
+                </osSpecific>
+            </osSpecifics>
+            <configuration-dependencies>
+                <config-type>zeppelin-ambari-config</config-type>
+                <config-type>zeppelin-config</config-type>
+            </configuration-dependencies>
+            <restartRequiredAfterChange>false</restartRequiredAfterChange>
+        </service>
+    </services>
+</metainfo>
