--- kerberos.json
@@ -1,17 +1,0 @@
-{
-  "services": [
-    {
-      "name": "ZEPPELIN",
-      "identities": [
-        {
-          "name": "/smokeuser"
-        }
-      ],
-      "components": [
-        {
-          "name": "ZEPPELIN_MASTER"
-        }
-      ]
-    }
-  ]
-}
