--- metainfo.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0"?>
-<metainfo>
-    <schemaVersion>2.0</schemaVersion>
-    <services>
-        <service>
-            <name>ZEPPELIN</name>
-            <displayName>Zeppelin Notebook</displayName>
-            <comment>A web-based notebook that enables interactive data analytics. It enables you to make beautiful data-driven, interactive and collaborative documents with SQL, Scala and more.</comment>
-            <version>0.0.5</version>
-            <components>
-                <component>
-                  <name>ZEPPELIN_MASTER</name>
-                  <displayName>Zeppelin Notebook</displayName>
-                  <category>MASTER</category>
-                  <cardinality>1</cardinality>
-                  <commandScript>
-                    <script>scripts/master.py</script>
-                    <scriptType>PYTHON</scriptType>
-                    <timeout>10000</timeout>
-                  </commandScript>
-                </component>
-            </components>
-            <osSpecifics>
-              <osSpecific>
-                <osFamily>redhat6</osFamily>
-                <packages>
-                   <package><name>nss</name></package>
-                   <package><name>git</name></package>
-                   <package><name>java-1.7.0-openjdk-devel</name></package>
-                   <package><name>apache-maven-3.2*</name></package>
-                </packages>
-              </osSpecific>
-              <osSpecific>
-                <osFamily>redhat7</osFamily>
-                <packages>
-                   <package><name>nss</name></package>
-                   <package><name>git</name></package>
-                   <package><name>java-1.7.0-openjdk-devel</name></package>
-                   <package><name>maven</name></package>
-                </packages>
-              </osSpecific>
-             <osSpecific>
-                <osFamily>ubuntu12</osFamily>
-                <packages>
-                   <package><name>git</name></package>
-                   <package><name>maven</name></package>
-                </packages>
-              </osSpecific>
-             <osSpecific>
-                <osFamily>ubuntu14</osFamily>
-                <packages>
-                   <package><name>git</name></package>
-                   <package><name>maven</name></package>
-                </packages>
-              </osSpecific>
-            </osSpecifics>
-      	    <configuration-dependencies>
-        	<config-type>zeppelin-ambari-config</config-type>
-        	<config-type>zeppelin-config</config-type>
-      	    </configuration-dependencies>
-            <restartRequiredAfterChange>false</restartRequiredAfterChange>
-        </service>
-    </services>
-</metainfo>
